handler functions fini

+++ repository/order/redis.go
@@ -16,8 +16,8 @@ type RedisRepo struct {
 	Client *redis.Client
 }
 type FindAllPage struct {
-	Size   uint
-	Offset uint
+	Size   uint64
+	Offset uint64
 }
 
 type FindResult struct {
